Added two things: soft shadows and anti-aliasing.
Soft Shadow: soft shading was implemented by averaging the shading of
several possible randomized paths to the light sources. This gives fairly
good soft shadows but at a high computational cost.

Anti-aliasing: used a multisampling approach: take 1, 4, or 16 evenly
distributed samples of a pixel. Naturally this adds to CPU time
dramatically; however, it lowers the required density of soft shadow
rays that need to be generated.

Signed-off-by: Alex Waterman <alex@australia.cpufarm.net>